pyproject.toml
[project]
name = "brain-unet-render"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "torch>=2.8.0",
]
